restyle: AlbumScreen, WebSearch & paginated infinite recommendations
